Error Code Scan Presentation
Properly added the error codes screen so that it initiates a scan and
can present results. The data provided at the moment it faked, returned
from methods not yet written.

The Detail panel has been added to show a PID or DTC in detail when
selected. It is brought up from clicking on a DTC, but as of yet does
not display anything.