Track Day & Graphing Added.
Added Track Day functions that time 1/4 mile and 0-60mph times, and
created managed leader boards that are also saved in their own files and
displayed on their own screen.

Have added the graph panel's contents so that PIDs can be checked. This
still needs some tweaking but has been tested live and showed that it
works! This required adding a widget so that graphs can be added to
dashboards too with as many PIDs as wanted.

Also created dialog to select PIDs from all available ones for the
current vehicle.