Changed some of the textures and appearances.
Updated some of the car system textures for polo 6n2 to show engine
outline, and added option in settings to change GUI colours. Also hoping
to be able to tint other textures like car ones to be able to change
them with the themes too.

Also added a bar dial widget, works mechanically, just sorting out
direction of text for PID title, and how it is displayed.